\(d_s\) for Between Subjects with Pooled SD Denominator

This function displays d for two between subjects groups and gives the central and non-central confidence interval using the pooled standard deviation as the denominator. Taken from MOTE for this package. Use MOTE for full options.

Usage

calculate_d(
  m1 = NULL,
  m2 = NULL,
  sd1 = NULL,
  sd2 = NULL,
  n1 = NULL,
  n2 = NULL,
  a = 0.05
)

Arguments

m1

mean group one

m2

mean group two

sd1

standard deviation group one

sd2

standard deviation group two

n1

sample size group one

n2

sample size group two

a

significance level

Value

Provides the effect size (Cohen's *d*) with associated central and non-central confidence intervals, the *t*-statistic, the confidence intervals associated with the means of each group, as well as the standard deviations and standard errors of the means for each group. The one-tailed confidence interval is also included for sensitivity analyses.

d

effect size

dlow

lower level confidence interval of d value

dhigh

upper level confidence interval of d value

Details

To calculate \(d_s\), mean two is subtracted from mean one and divided by the pooled standard deviation. $$d_s = \frac{M_1 - M_2}{S_{pooled}}$$

Examples

calculate_d(m1 = 4, m2 = 3,
sd1 = 1, sd2 = 1,
n1 = 100, n2 = 100,
a = .05)
#> $d
#> [1] 1
#> 
#> $dlow
#> [1] 0.7047339
#> 
#> $dhigh
#> [1] 1.293011
#>