Calculation of Partial Invariance by Parameter

This function iterates over a constrained model and examines potential areas for partial invariance. The function returns a list of models and a dataframe so that the user can decide which items may be relaxed for partial invariance.

Usage

partial_mi(saved_model, data, model, group, group.equal, partial_step, ...)

Arguments

saved_model

A lavaan model with at least one level of constraints across groups.

data

Dataframe from the original saved model.

model

Model from the original saved model.

group

Grouping variable column from the original saved model. Use a character vector.

group.equal

A vector of names of the constraints to test for multigroup analysis, in the order that they should be tested. See lavaan for possible options.

partial_step

The level of partial invariance you would like to test. You can use "loadings", "regressions", "intercepts", "residuals", or "thresholds". Note that for some models this may estimate more than you want (i.e., the syntax for factor covariances is ~~ as well as residuals) but these can be excluded from the from the final dataframe.

...

Any other lavaan arguments you would like to add.

Value

Models with each constraint relaxed individually and a summary table of fit indices from each model.

models

Saved lavaan models for each parameter relaxed.

fit_table

A dataframe of fit indices for all models to investigate places for partial invariance.

Examples

HS.model <- ' visual  =~ x1 + x2 + x3
textual =~ x4 + x5 + x6
speed   =~ x7 + x8 + x9 '

library(lavaan)

data("HolzingerSwineford1939")

saved_mgcfa <- mgcfa(model = HS.model,
 data = HolzingerSwineford1939,
 group = "sex",
 group.equal = c("loadings", "intercepts", "residuals"),
 meanstructure = TRUE)

saved_mi <- partial_mi(saved_model =
  saved_mgcfa$invariance_models$model.loadings,
  data = HolzingerSwineford1939,
  model = HS.model,
  group = "sex",
  group.equal = "loadings",
  partial_step = "loadings")

  saved_mi$fit_table
#> # A tibble: 9 × 47
#>   free.parameter npar       fmin      chisq    df    pvalue       baseline.chisq
#>   <chr>          <lvn.vctr> <lvn.vct> <lvn.vc> <lvn> <lvn.vctr>   <lvn.vctr>    
#> 1 visual =~ x1   54         0.2096778 126.2260 54    1.025522e-07 967.2277      
#> 2 visual =~ x2   55         0.2079012 125.1565 53    9.021992e-08 967.2277      
#> 3 visual =~ x3   55         0.2046712 123.2120 53    1.616199e-07 967.2277      
#> 4 textual =~ x4  54         0.2096778 126.2260 54    1.025522e-07 967.2277      
#> 5 textual =~ x5  55         0.2087596 125.6733 53    7.719583e-08 967.2277      
#> 6 textual =~ x6  55         0.2094712 126.1017 53    6.781510e-08 967.2277      
#> 7 speed =~ x7    54         0.2096778 126.2260 54    1.025522e-07 967.2277      
#> 8 speed =~ x8    55         0.1956719 117.7945 53    7.944429e-07 967.2277      
#> 9 speed =~ x9    55         0.1829277 110.1225 53    6.944395e-06 967.2277      
#> # ℹ 40 more variables: baseline.df <lvn.vctr>, baseline.pvalue <lvn.vctr>,
#> #   cfi <lvn.vctr>, tli <lvn.vctr>, nnfi <lvn.vctr>, rfi <lvn.vctr>,
#> #   nfi <lvn.vctr>, pnfi <lvn.vctr>, ifi <lvn.vctr>, rni <lvn.vctr>,
#> #   logl <lvn.vctr>, unrestricted.logl <lvn.vctr>, aic <lvn.vctr>,
#> #   bic <lvn.vctr>, ntotal <lvn.vctr>, bic2 <lvn.vctr>, rmsea <lvn.vctr>,
#> #   rmsea.ci.lower <lvn.vctr>, rmsea.ci.upper <lvn.vctr>,
#> #   rmsea.ci.level <lvn.vctr>, rmsea.pvalue <lvn.vctr>, …
  saved_mi$models
#> $`visual =~ x1`
#> lavaan 0.6-19 ended normally after 47 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of model parameters                        60
#>   Number of equality constraints                     6
#> 
#>   Number of observations per group:                   
#>     1                                              146
#>     2                                              155
#> 
#> Model Test User Model:
#>                                                       
#>   Test statistic                               126.226
#>   Degrees of freedom                                54
#>   P-value (Chi-square)                           0.000
#>   Test statistic for each group:
#>     1                                           72.534
#>     2                                           53.692
#> 
#> $`visual =~ x2`
#> lavaan 0.6-19 ended normally after 49 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of model parameters                        60
#>   Number of equality constraints                     5
#> 
#>   Number of observations per group:                   
#>     1                                              146
#>     2                                              155
#> 
#> Model Test User Model:
#>                                                       
#>   Test statistic                               125.157
#>   Degrees of freedom                                53
#>   P-value (Chi-square)                           0.000
#>   Test statistic for each group:
#>     1                                           71.454
#>     2                                           53.702
#> 
#> $`visual =~ x3`
#> lavaan 0.6-19 ended normally after 54 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of model parameters                        60
#>   Number of equality constraints                     5
#> 
#>   Number of observations per group:                   
#>     1                                              146
#>     2                                              155
#> 
#> Model Test User Model:
#>                                                       
#>   Test statistic                               123.212
#>   Degrees of freedom                                53
#>   P-value (Chi-square)                           0.000
#>   Test statistic for each group:
#>     1                                           71.369
#>     2                                           51.843
#> 
#> $`textual =~ x4`
#> lavaan 0.6-19 ended normally after 47 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of model parameters                        60
#>   Number of equality constraints                     6
#> 
#>   Number of observations per group:                   
#>     1                                              146
#>     2                                              155
#> 
#> Model Test User Model:
#>                                                       
#>   Test statistic                               126.226
#>   Degrees of freedom                                54
#>   P-value (Chi-square)                           0.000
#>   Test statistic for each group:
#>     1                                           72.534
#>     2                                           53.692
#> 
#> $`textual =~ x5`
#> lavaan 0.6-19 ended normally after 47 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of model parameters                        60
#>   Number of equality constraints                     5
#> 
#>   Number of observations per group:                   
#>     1                                              146
#>     2                                              155
#> 
#> Model Test User Model:
#>                                                       
#>   Test statistic                               125.673
#>   Degrees of freedom                                53
#>   P-value (Chi-square)                           0.000
#>   Test statistic for each group:
#>     1                                           72.201
#>     2                                           53.472
#> 
#> $`textual =~ x6`
#> lavaan 0.6-19 ended normally after 49 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of model parameters                        60
#>   Number of equality constraints                     5
#> 
#>   Number of observations per group:                   
#>     1                                              146
#>     2                                              155
#> 
#> Model Test User Model:
#>                                                       
#>   Test statistic                               126.102
#>   Degrees of freedom                                53
#>   P-value (Chi-square)                           0.000
#>   Test statistic for each group:
#>     1                                           72.455
#>     2                                           53.647
#> 
#> $`speed =~ x7`
#> lavaan 0.6-19 ended normally after 47 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of model parameters                        60
#>   Number of equality constraints                     6
#> 
#>   Number of observations per group:                   
#>     1                                              146
#>     2                                              155
#> 
#> Model Test User Model:
#>                                                       
#>   Test statistic                               126.226
#>   Degrees of freedom                                54
#>   P-value (Chi-square)                           0.000
#>   Test statistic for each group:
#>     1                                           72.534
#>     2                                           53.692
#> 
#> $`speed =~ x8`
#> lavaan 0.6-19 ended normally after 48 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of model parameters                        60
#>   Number of equality constraints                     5
#> 
#>   Number of observations per group:                   
#>     1                                              146
#>     2                                              155
#> 
#> Model Test User Model:
#>                                                       
#>   Test statistic                               117.794
#>   Degrees of freedom                                53
#>   P-value (Chi-square)                           0.000
#>   Test statistic for each group:
#>     1                                           67.657
#>     2                                           50.138
#> 
#> $`speed =~ x9`
#> lavaan 0.6-19 ended normally after 49 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of model parameters                        60
#>   Number of equality constraints                     5
#> 
#>   Number of observations per group:                   
#>     1                                              146
#>     2                                              155
#> 
#> Model Test User Model:
#>                                                       
#>   Test statistic                               110.122
#>   Degrees of freedom                                53
#>   P-value (Chi-square)                           0.000
#>   Test statistic for each group:
#>     1                                           63.775
#>     2                                           46.347
#>