Visualization for Measurement Invariance

A function to create visualizations for measurement invariance comparison between two groups. Please note: this function graphs two groups at a time - use the function several times if you have more than two groups.

Usage

plot_mi(
  data_coef,
  model_step,
  item_name,
  x_limits = c(-1, 1),
  y_limits,
  conf.level = 0.95,
  model_results,
  lv_name,
  plot_groups = NULL
)

Arguments

data_coef

A tidy dataframe of the coefficients from mgcfa function in this package.

model_step

The model step you would like to plot using this function. Check the "Model" column in the coefficients to pick. Include this as a character vector.

item_name

The name of the item/variable you want to plot for comparison. Check out the "term" column to know your options. Usually observed variables from your model.

x_limits

The limits of the latent variable to plot on on the latent mean graph.

y_limits

The limits of the observed variable that you included in item_name.

conf.level

The confidence interval level you would like to graph.

model_results

The saved summary from cfa or mgcfa of the same model name you listed in model_step.

lv_name

The name of the latent variable you would like to plot. Ensure this name is the same as what is in the data_coef.

plot_groups

An optional variable to denote which two groups you would like to plot. Used when more than two groups are included in the output.

Value

A list of graphs created to visualize measurement invariance.

complete

A ggplot2 object that includes all visualization stacked together for publication.

intercept

A ggplot2 object that includes the intercept section of the overall graph.

mean

A ggplot2 object that includes the latent means section of the overall graph.

variance

A ggplot2 object that includes the variance/residuals section of the overall graph.

Examples

HS.model <- ' visual  =~ x1 + x2 + x3
textual =~ x4 + x5 + x6
speed   =~ x7 + x8 + x9 '

library(lavaan)

data("HolzingerSwineford1939")

saved_model <- cfa(HS.model, data = HolzingerSwineford1939,
 meanstructure = TRUE,
 group = "sex",
 group.equal = c("loadings"))

 saved_mgcfa <- mgcfa(model = HS.model,
  data = HolzingerSwineford1939,
  group = "sex",
  group.equal = c("loadings", "intercepts", "residuals"),
  meanstructure = TRUE)

# saved_mi_plots <- plot_mi(data_coef = saved_mgcfa$model_coef,
#  model_step = "Configural", # which model
#  item_name = "x1", # name of observed item
#  x_limits = c(-1,1), # LV limits to graph
#  y_limits = c(min(HolzingerSwineford1939$x1),
#             max(HolzingerSwineford1939$x1)), # Y min and max in data
#  conf.level = .95, # what ci do you want
#  model_results = saved_mgcfa$model_configural, # what model results do you want
#  lv_name = "visual", # which latent is the observed variable on
#  plot_groups = NULL)