Bootstrapping Measurement Invariance

A function to help bootstrap expected rates of "replication" for measurement invariance as compared to a random assignment of groups. This function is mostly designed for estimation for registered reports but can also be used after a model is completed to estimate rates of expected differences in the future.

Usage

bootstrap_model(
  saved_configural,
  data,
  model,
  group,
  nboot = 1000,
  invariance_index,
  invariance_rule,
  group.equal,
  ...
)

Arguments

saved_configural

A saved lavaan model of the "configural" model that includes the grouping variable but no other equality constraints.

data

The dataframe for the estimation

model

The original model lavaan syntax

group

The grouping variable column as a character

nboot

The number of bootstraps you would like to calculate. Please note: large models with many parameters will run slowly depending on your computer.

invariance_index

The name of the fix index you want to use for invariance testing. For example, "cfi", or "rmsea" - must be a character vector and part of the fitmeasures provided by lavaan.

invariance_rule

The difference between your previous model fit index and the new equality model index that you would accept as invariant. Must be a numeric value.

group.equal

The equality constraints you would like to impose as a vector, in order. This argument is the same as lavaan syntax. For example, c("loadings", "intercepts", "residuals").

...

Other arguments to be included to configure the cfa function from lavaan. For example, you can include arguments for ordered models, clustering, sampling.weights, or estimators. Note: the model is updated with this information, so if you had it in the mgcfa model, no need to specify it again.

Value

A dataframe of the bootstrapped results. This set includes the proportion of invariant tests for bootstrapping and random results and the effect size h of this difference.

Examples

HS.model <- ' visual  =~ x1 + x2 + x3
textual =~ x4 + x5 + x6
speed   =~ x7 + x8 + x9 '

library(lavaan)
#> This is lavaan 0.6-19
#> lavaan is FREE software! Please report any bugs.

data("HolzingerSwineford1939")

saved_model <- cfa(HS.model, data = HolzingerSwineford1939,
 meanstructure = TRUE,
 group = "sex",
 group.equal = c("loadings"))

saved_configural <- cfa(HS.model, data = HolzingerSwineford1939,
 meanstructure = TRUE,
 group = "sex")

# saved_boot <- bootstrap_model(
#  saved_configural = saved_configural,
#  data = HolzingerSwineford1939,
#  model = HS.model,
#  group = "sex",
#  nboot = 100, # don't do this - this is to make it run fast
#  invariance_index = "cfi",
#  invariance_rule = .01,
#  group.equal = c("loadings", "intercepts", "residuals"))

#  saved_boot